Subclass additional User functionality to ExtendedIRCUser class
This is a breaking change. The functionality of IRCUser and IRCUserSnapshot's getName() method has been changed to return the nick instead of user mask. The user mask is now found under IRCUserSnapshot.getMask()
IRCUserSnapshot's fields are also no longer guaranteed to not be null, it now depends on if the snapshot is generated from an IRCUser or an ExtendedIRCUser.