[BREAKING] Change Channel interface, support NAMES
This commit should provide a much better solution to the problem that is
rate limiting of the WHO command preventing useful information such as who
is OP in a channel. When the client joins a channel it is immediately
sent, by the server, a full NAMES output. This should happen prior to
anything else important happening in the channel and such is a great
opportunity for determining modes.

Thus, I have separated the concept of User from modes, which is now tied
only to a String nick value. Nicks are highly useful for so much, there's
no reason to wait until the client acquires full User details.