Added simple .gitignore:
This file only ignores the /target directory as it is the only thing that
will exist and not ignored by a default clone and build. Files that only
exist for one OS or IDE are not listed to avoid having to update this
every time an IDE or OS changes plans.

See https://help.github.com/articles/ignoring-files/ for more info on how
you can exclude things on a system or repo-specific level.