Do not untrack channel on other user's part
Currently, when a different user (as in, not the user this client is connected as) parts a channel, that channel will be untracked, just as if you left the channel. Thus subsequent messages / parts will cause issues as the client thinks it just left the channel. This doesn't seem to be the correct behavior, as we would only want to invalidate the user that parted the channel instead of leaving it.