Add twitch whisper event
This adds the whisper event for whisper messages sent via twitch.
Whisper messages are done via the twitch command "WHISPER" and have the
target (i.e. the client this socket is connected as) and the message as
parameters. Sending a whisper works a little different from normal IRC
as well. You have to send a message to the "#jtv" channel with the
format "/w <target_username> <message>".